backend/requirements.txt
# Backend Bağımlılıkları
# =====================
# Kurulum: pip install -r requirements.txt

# FastAPI - Web framework
fastapi

# Uvicorn - ASGI server (FastAPI'yi çalıştırmak için)
uvicorn[standard]

# LangChain - LLM framework
langchain
langchain-core
langchain-openai
langgraph

# Pydantic
pydantic
pydantic-settings

# PostgreSQL - Veritabanı bağlantısı
psycopg2-binary

# Redis response cache
redis

# Async HTTP for MCP adapters
httpx
mcp
tavily-python
python-multipart

# Observability
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc

# Local dev - .env loading
python-dotenv

# Testing
pytest
